Module Olm__Pk

This module contains bindings to the PK part of the Olm library. It contains the modules Decryption and Encryption that are used to establish an encrypted communication channel using public key encryption, as well as the module Signing that is used to sign a message.

module Message : sig ... end

module Encryption : sig ... end

module Decryption : sig ... end

module Signing : sig ... end